requirements.txt
fastapi>=0.95.0,<1.0.0
uvicorn>=0.22.0,<1.0.0
pydantic>=1.10.8,<2.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.26.0  
prophet>=1.1.0,<2.0.0
sqlalchemy>=2.0.0,<3.0.0
pymysql>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pyjwt>=2.7.0,<3.0.0
python-multipart>=0.0.6,<0.1.0
joblib>=1.2.0,<2.0.0
setuptools>=69.0.0  
